requirements.txt
geopandas
psycopg2
geoalchemy2
osmnx
pandana
tqdm
python-dotenv
mkdocs-material
mkdocstrings
markdown-include
rich
pg-data-etl
git+https://github.com/aaronfraint/postgis-helpers.git
git+https://github.com/aaronfraint/philly-transit-data.git